Added Floor Organization and Sprites for Scenes
I added folders for the scenes for the different floors, they are currently empty. I worked on my testing area and got a decent movement system working, it sticks to the floor when trying to jump from a dead stop so I'm still working on it. Added some free sprite assets from the asset store to use as placeholders for our scenes if we can't find better ones.